moved al logic related to managing client state into a new owa_client class. all php clients now derive from owa_client cleaned up handlers by removing checks for inbound session/visitor id ading missing queue_item entity moved tracking methods from owa_caller to owa_client

+++ modules/base/classes/state.php
@@ -141,13 +141,15 @@ class owa_state {
 		
 	function clear($store) {
 		
-		unset($this->stores[$store]);
-		
-		if ($this->stores_meta[$store]['type'] === 'cookie') {
-		
-			return owa_coreAPI::deleteCookie($store);	
-		}	
-		
+		if (array_key_exists($store, $this->stores)) {
+			
+			unset($this->stores[$store]);
+			
+			if ($this->stores_meta[$store]['type'] === 'cookie') {
+			
+				return owa_coreAPI::deleteCookie($store);	
+			}	
+		}		
 	}
 	
 	function getPermExpiration() {
